requirements.txt
pillow>=10.0.0
pydicom>=2.4.0
numpy>=1.24.0
requests>=2.31.0
python-magic-bin>=0.4.14
ttkthemes>=3.2.2
tk>=0.1.0
